+++ R/Wrapper_MixtureDistribution.R
@@ -2,13 +2,14 @@
 #' @title Mixture Distribution Wrapper
 #' @description Wrapper used to construct a mixture of two or more distributions.
 #'
-#' @section Constructor: MixtureDistribution$new(distlist, weights = NULL)
+#' @section Constructor: MixtureDistribution$new(distlist, weights = NULL, vectordist = NULL)
 #'
 #' @section Constructor Arguments:
 #' \tabular{lll}{
 #' \strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
 #' \code{distlist} \tab list \tab List of distributions. \cr
 #' \code{weights} \tab numeric \tab Vector of weights. See Details. \cr
+#' \code{vectordist} \tab numeric \tab Vector Distribution. See Details. \cr
 #' }
 #'
 #' @details A Mixture Distribution is a weighted combination of two or more distributions such that for
@@ -21,6 +22,9 @@
 #' If weights are given, they should be provided as a vector of numerics. If they don't sum to one
 #' then they are normalised automatically. If NULL, they are taken to be uniform, i.e. for n
 #' distributions, \eqn{w_i = 1/n, \ \forall \ i \ \in \ [1,n]}{w_i = 1/n, for all i \epsilon [1,n]}.
+#'
+#' Can optionally be constructed using a \code{VectorDistribution}, in which case \code{distlist} is ignored
+#' and the mixture is constructed with the wrapped models in the vector.
 #'
 #'
 #' @inheritSection DistributionWrapper Public Variables
@@ -39,9 +43,18 @@
 #' @export
 NULL
 MixtureDistribution <- R6::R6Class("MixtureDistribution", inherit = DistributionWrapper, lock_objects = FALSE)
-MixtureDistribution$set("public","initialize",function(distlist, weights = NULL){
+MixtureDistribution$set("public","initialize",function(distlist, weights = NULL, vectordist = NULL){
 
-  distlist = makeUniqueDistributions(distlist)
+  name = short_name = description = NULL
+
+  if(!is.null(vectordist)){
+    distlist <- vectordist$wrappedModels()
+    name = gsub("Vector","Mixture",vectordist$name)
+    short_name = gsub("Vec","Mix",vectordist$short_name)
+    description = gsub("Vector","Mixture",vectordist$description)
+  } else
+    distlist <- makeUniqueDistributions(distlist)
+
   distnames = names(distlist)
 
   if(is.null(weights))
@@ -74,16 +87,19 @@ MixtureDistribution$set("public","initialize",function(distlist, weights = NULL)
     y = c()
     for(i in 1:length(x))
       y = c(y, self$wrappedModels()[[i]]$rand(x[[i]]))
-    return(sample(y, n))
+    if(length(y) == 1)
+      return(y)
+    else
+      return(sample(y, n))
   }
   formals(rand)$self <- self
 
-  name = paste("Mixture of",paste(distnames, collapse = " and "))
-  short_name = paste(distnames, collapse = "Mix")
+  if(is.null(name)) name = paste("Mixture of",paste(distnames, collapse = " and "))
+  if(is.null(short_name)) short_name = paste(distnames, collapse = "Mix")
   type = do.call(union.SetInterval, lapply(distlist, type))
   support = do.call(union.SetInterval, lapply(distlist, type))
 
-  description =  paste0("Mixture of: ",paste0(1:length(distlist),") ",lapply(distlist, function(x) x$description),
+  if(is.null(description)) description =  paste0("Mixture of: ",paste0(1:length(distlist),") ",lapply(distlist, function(x) x$description),
                                             collapse = " And "), " - With weights: (",
                        paste0(weights, collapse=", "), ")")