+++ NEWS.md
@@ -14,10 +14,11 @@
 
 - Added quantile to Categorical distribution and updated its cdf efficiency
 - `cdf` of discrete distributions evaluated between integers are now evaluated after rounding down and do not return 0, e.g. `Binomial$new()$cdf(1.8) == Binomial$new()$cdf(1)`
-- Added two additional parameters to `ProductDistribution`: `paramList` and `distribution`, so that it can either function as before or as the `ArrayDistribution` wrapper which is now deprecated.
-- Added two additional parameters to `VectorDistribution`: `paramList` and `distribution`, so that it can either function as before or as a generalisation to the `ArrayDistribution` wrapper which is now deprecated.
+- Added two additional parameters to `ProductDistribution`: `params` and `distribution`, so that it can either function as before or as the `ArrayDistribution` wrapper which is now deprecated.
+- Added two additional parameters to `VectorDistribution`: `params` and `distribution`, so that it can either function as before or as a generalisation to the `ArrayDistribution` wrapper which is now deprecated.
 - Removed automated rounding of numerics when `ParameterSet` expected integer in `setParameterValue`
 - Added parameter `n` to `strprint` and `print` to clean the print method for distributions (especially wrappers) with multiple parameters
+- Added optional `vectordist` argument to `MixtureDistribution` for better compatibility with `VectorDistribution`
 
 ## Patches
 
@@ -25,6 +26,8 @@
 - Fixed the `rand` return of a Vector Distribution (transposed data.table and added column names)
 - Fixed spelling mistakes in extension guidelines and removed calls to redundant parameters 
 - Removed redundant `crayon` dependency
+- Fixed bug in `MixtureDistribution` that was causing `rand` to erroneously return integers
+- `Degenerate` distribution re-classified as discrete
 
 # distr6 1.0.1